# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: semochal API 기능 명세서
  version: 0.0.1
  description: |
    챌린지(공모전) 매칭 플랫폼 semochal의 기능 명세서.

    - **사용자 영역**: 홈/탐색, 팀 모집, 챌린지 지원, 마이페이지(북마크·관심분야·알림·프로필)
    - **기업 영역(SEMO.BIZ)**: 기관 가입/인증, 공고(챌린지) 관리, 지원서 관리, 광고/결제, 운영대행 문의
    - **관리자 영역**: 기관 심사, 인증 관리, 광고 관리/광고비, 사용자 관리, 신고 처리, 리포트, 설정

    구현 상태 표기:
    - `implemented` — `server/src/modules`의 컨트롤러에 실제 구현된 엔드포인트
    - `planned` — 프론트 화면(`front/src/app`)에서 사용되지만 아직 백엔드 미구현(목업 데이터 기반)

    인증 방식: 로그인 시 설정되는 HttpOnly 세션 쿠키의 JWT를 사용한다.
servers:
  - url: http://localhost:3001
    description: 로컬 개발 서버
...
paths:
...
  /ads:
    get:
      tags: [Ads]
      summary: 내 광고 목록 조회
      operationId: listMyAds
      x-impl-status: implemented
      description: 광고 관리 화면(`/biz/ads`)의 진행중인 광고 테이블과 대시보드 차트용.
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [active, paused, ended] }
      responses:
        '200':
          description: 광고 목록
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Ad' } }
    post:
      tags: [Ads]
      summary: 광고 등록 (기간 선택 + 결제)
      operationId: createAd
      x-impl-status: implemented
      description: |
        단기 결제 팝업(react-day-picker 기간 선택, 일일 단가 × 기간 = 결제 금액) → 결제수단 확인 → 완료.
        주문 생성(`orders.adId`) → 토스 결제 → 웹훅(`POST /payments/webhook/toss`)이
        주문을 paid로 바꾸면 `OrdersService.markPaid`가 이 광고를 자동으로 `active`로 전환한다.
        생성 직후에는 결제 대기 상태인 `preparing`으로 반환된다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, startDate, endDate]
              properties:
                productId: { type: string }
                expectedDailyPrice:
                  {
                    type: integer,
                    minimum: 0,
                    description: 화면에서 확인한 단가. 최신 단가와 다르면 409로 재확인 요청.,
                  }
                imageFileId: { type: string, format: uuid }
                landingUrl: { type: string }
                startDate: { type: string, format: date }
                endDate: { type: string, format: date }
                title:
                  {
                    type: string,
                    description: 'implemented 확장 — 원 스키마엔 title/challengeId가 없어 Ad.title의 출처가 비어 있었음. 생략 시 "{상품명} 광고"로 대체.',
                  }
      responses:
        '201':
          description: 생성된 광고 (결제 대기). 계약과 주문 금액은 생성 시점 단가로 고정. 종료일 포함 일 단위 계약이며 기존 계약 종료일 다음 날부터 예약 가능.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ad' }
        '409':
          description: 기존 계약과 기간 중복 또는 확인한 단가가 변경됨
...
  /admin/businesses:
    get:
      tags: [Admin]
      summary: 기관 심사 목록 + 심사 현황 통계 조회
      operationId: listAdminBusinesses
      x-impl-status: implemented
      description: |
        기관 심사(`/admin/biz-review`).
        - `stats`: 상단 통계 카드 4종(전체 신청/승인/거부/대기) — dot 색상 포함
        - `items`: 기관 테이블 — 활동유형(org)/유형/사업자번호(마스킹)/신청일/NTS OCR 결과/상태 뱃지.
          행 클릭 시 우측 상세 패널(기관 정보 + 대기 상태일 때 승인/거부 버튼) 표시.
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema: { type: string, description: 기관 유형 (비영리/학교/협회/기업) }
        - name: status
          in: query
          schema: { type: string, enum: [pending, approved, rejected] }
      responses:
        '200':
          description: 기관 심사 목록과 통계
          content:
            application/json:
              schema:
                type: object
                properties:
                  stats: { $ref: '#/components/schemas/AdminStatCardList' }
                  items: { type: array, items: { $ref: '#/components/schemas/BizReviewEntry' } }
...
  /admin/ads:
    get:
      tags: [Admin]
      summary: 전체 광고 목록 조회
      operationId: listAdminAds
      x-impl-status: implemented
      description: '광고 관리(`/admin/ad-pricing`): 검색/상태/활동 필터, 광고 수정·납볍·중단.'
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200':
          description: 광고 목록
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Ad' } }
  /admin/ad-pricing:
    get:
      tags: [Admin]
      summary: 광고 단가 조회
      operationId: getAdPricing
      x-impl-status: implemented
      description: |
        광고비 관리(`/admin/ad-pricing/pricing`): PC/모바일 뷰 토글로 홈 화면 미리보기(MovingAds 캐러셀,
        챌린지/팀 카드)를 재현. 슬롯 클릭 → 다이얼로그(현재 광고사/기간, '리포트 보기' → `/admin/analytics?ad=N`)
        → 금액 수정하기에서 하루 광고비 입력/저장.
      responses:
        '200':
          description: 슬롯별 단가 정보
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdSlotPricing' }
    put:
      tags: [Admin]
      summary: 광고 단가 수정
      operationId: updateAdPricing
      x-impl-status: implemented
      description: 슬롯 클릭 다이얼로그에서 하루 광고비 입력/저장 (`adPricing.dailyPrice`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                required: [slot, dailyPrice]
                properties:
                  slot: { type: string, enum: [hero, gallery, team] }
                  dailyPrice: { type: integer, description: 일일 광고비 (원) }
      responses:
        '200':
          description: 저장된 단가
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/AdSlotPricing' } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT 액세스 토큰(향후 도입). 현재는 `x-user-id` 헤더로 사용자를 식별하는 개발 중 상태.
...
  schemas:
...
    Ad:
      type: object
      properties:
        id: { type: string, format: uuid }
        businessId: { type: string, format: uuid }
        productId: { type: string }
        title: { type: string, description: 광고명 (연결된 챌린지명) }
        imageFileId: { type: string, format: uuid }
        landingUrl: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        status:
          {
            type: string,
            enum: [active, preparing, paused, ended],
            description: 진행중/준비중/중단됨/만료 — 대시보드 '진행중인 광고' 테이블 표기용,
          }
        impressions: { type: integer, nullable: true, description: 누적 노출수 (준비중이면 null) }
        bookmarks: { type: integer, nullable: true, description: 누적 북마크수 (준비중이면 null) }
        paidAmount: { type: integer, description: 결제 금액 (원) }
...
    AdminStatCard:
      type: object
      description: |
        관리자 화면 통계 카드 (StatRow/StatCard) — label + 포맷된 값 + 보조 문구 + 도트 색.
        서버는 value를 원시 수치로, 포맷('1,248'·'₩42.8M')은 화면에서 처리하는 것을 권장.
      properties:
        label: { type: string, description: '카드 제목 (예: 승인된 기관)' }
        value: { type: string, description: '표시 값 (예: 1,248 / ₩42.8M)' }
        meta: { type: string, description: '보조 문구 (예: +12 이번 주 / 월간)' }
        dot: { type: string, nullable: true, description: '도트 색상 (hex, 예: #0877FF)' }
    AdminStatCardList:
      type: array
      items: { $ref: '#/components/schemas/AdminStatCard' }
...
    BizReviewEntry:
      type: object
      description: 기관 심사 테이블 행 (BizRow). 사업자번호는 서버에서 마스킹 후 전달.
      properties:
        id: { type: string, format: uuid }
        org: { type: string, description: 기관명 (활동유형 컬럼) }
        type: { type: string, description: 기관 유형 (비영리/학교/협회/기업) }
        bizNumber: { type: string, description: '마스킹된 사업자번호 (예: 110-81-*****)' }
        appliedAt: { type: string, description: '신청일 (예: 05.12)' }
        nts:
          {
            type: string,
            enum: [success, failed, closed, unrecognized],
            description: 'NTS OCR 검증 결과 — 성공/실패/폐업·폐점/인식불가 (실패 계열은 빨간색 표기)',
          }
        status:
          {
            type: string,
            enum: [pending, approved, rejected],
            description: 심사 상태 — 대기(파랑)/승인(초록)/거부(빨강) 뱃지,
          }
...
    AdSlotPricing:
      type: object
      description: 광고비 관리 슬롯 항목 — 슬롯별 단가와 현재 광고 정보
      properties:
        slot: { type: string, enum: [hero, gallery, team] }
        dailyPrice: { type: integer, description: 일일 광고비 (원) }
        currentAdId: { type: string, format: uuid, nullable: true }
        organization: { type: string, nullable: true, description: '현재 광고사 (예: 부산광역시)' }
        period: { type: string, nullable: true, description: "노출 기간 (예: '8/24~9/24')" }
...